Telnet Connection: Fix Telnet connection
Telnet Connection class contains a bug that make it unusable.
It was not possible to communicate with the target via telnet and via serial
console at the same time. This patch enables opening multiple telnet
connections to the target in parallel.

The following files were modified:
* source/target/equipment_connection.rb
Only make telnel connection the default if there is no connection already established

* source/target/telnet_equipment_connection.rb
Re-enable code that was mistakenly commented out and it is required for proper
operation of send_cmd method

Signed-off-by: Carlos Hernandez <ceh@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>