Add automatic check of performance data
Currently it is difficult to figure out if the performance results
obtained on a given release/platform are worst than results obtained
on previous releases.

This patch automatically compares the performance results with the
best performance obtained for the same platform and it forces a
test failure if performance is degraded by more than 5%.

A new module PassCriteria queries a new web service, which is not
part of VATF and which was created to support this functionality.
The web service returns the max, min and avg values for the requested
(platform, testcase, performance metric) tuple and then the
PassCriteria module compares the performance data with the data
returned by the web service.

At the moment, most of the comparisons are done using MAX value.
There are few exceptions like cpuload, latency and other performance
metrics were the comparison is done using MIN value.
The current implementation does not use AVG value due to 'bad' zero
samples stored in the performance table for a significant number of
tests.

Users can disable or tune the automatic check functionality from
their vatf-scripts. In particular, users can:
1) skip the automatic check functionality by defining
   a method named 'skip_perf_comparison'. The method can be no op.
2) overwrite the comparison operator 'min', 'avg' or 'max' by defining
   method overwrite_perf_comparison_operator(testcase_id, metric_name)
3) overwrite the data value to compare against by defining
   overwrite_perf_value(platform, testcase_id, metric_name, op, data)

Signed-off-by: Carlos Hernandez <ceh@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>