Reduce boot wait time from 10min to 3 min
This is required to reduce total test session time (specially on
nightly builds) where kernel builds are particularly unstable.
The 10-minute wait time was set because amsdk builds used to take
a long time to boot on a fresh filesystem. However, it is believe
than 3 minutes should be enough time to boot even a fresh fs.

Signed-off-by: Carlos Hernandez <ceh@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>