SystemLoader: Reduce boot timeout from 15 to 5 min
Some time ago, our filesystems used to have lengthly boot times, so
we setup a genereous boot timeout of 15 minutes. The problem with
such long timeout is that it extends test time considerably when
the board is not even booting. Moreover, current filesystem
typically boot in much less than 5 minutes, so a 5-minute timeout
should be more than enough.

Signed-off-by: Carlos Hernandez <ceh@ti.com>