Patch to include an array of power ports for one DUT
This patch makes changes to power_handler.rb to take
care of an array at dut.power_port element of bench file.
The power ports will be reset in the order in which
they appear in the array.
Also included is a sample of bench entry for such a setup.
No change is required in existing bench entries if only
one power port entry is required.

V2 - fixed indentation

Signed-off-by: Aparna Balasubramanian <aparnab@ti.com>
Signed-off-by: Carlos Hernandez <ceh@ti.com>