Add wxruby based windowing library
This patch adds a windowing library based on wxruby (32 bits systems)
and wxruby-ruby19 (64 bit systems). This library allows the user add
a results window to test scripts which may be useful in semi-automated
test cases. The patch also removes the old windowing library that was
based on rubyclr which only worked on Windows OS.

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>