Refactor atf_session_runner to isolate tests from execution engine
This patch refactors atf_session_runner so that test cases are ran
in a separate process from the execution engine. By running the
test cases in a separate process the execution engine is isolated
from any changes the tests may try to perform in shared data. The
patch also reverts changes made trying to protect shared data and
cleanup of files required by the test cases, since those changes
are not required after this patch.

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>