Refactor Devantech realy driver
This patch refactors the Devantech relay driver so that the
switch_off/switch_on functions comply with the behavior of
other power controllers like APC. Before this patch switch_on
would open the NC contacts that remove power from the boards
connected to them and switch_off closed them which restored
power. The way the other power controllers implemented the
functions switch_off removes power and switch_on restores power.
This patch refactors the behavior of those functions in the
Devantech driver to work like the other power controller drivers,
which allows the use of this and other power controller drivers
in scripts that call the switch_off and switch_on function directly

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>