linux_equipment_driver: Fix serial disconnect call on power cycle
This patch fixes the disconnect call on the power_cycle function.
This function was calling disconnect with a Hash parameter and the
function definition of disconnect takes a string as a parameter.
This patch fixes the problem and adds a wait for the tty dev node
before proceeding to reopen a serial connection to the dut

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Carlos Hernandez <ceh@ti.com>