LinuxEquipmentDriver: Remove noinitrd from bootargs
This patch remove noinitrd from the default bootargs of the boards
in the LinuxEquipmentDriver class. The noinitrd entry in bootargs
prevents booting from a ramdisk fs which is a use case on Linux
releases. After this patch the boards that had noinitrd in their
default bootargs are able to boot with and ramdisk fs

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>