LinuxEquipmentDriver: Try to power cycle if no serial connection
This patch changes the power_cycle function so that it will try
to power cycle the board even if the attempt to connect to the
board using the serial port fails. Before this patch if
the connect attempt failed the driver would raise an exception.
After this patch if the serial connection attempt fails the
driver will still try to power cycle the board for bench entries
that have the power_port property specified. This functionality
helps for boards like beaglebone, dra7xx, etc that have an ftdi
chip for serial connectivity and may be connected but have been
powered off.

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>