Update to fix a nil error condition in set_systemloader
This update fixes the set_systemloader method so that if
params['boot_cmds'] equals nil it will still properly match without
causing an exception. Also simplified the other elseif lines since
there is no need to check if the hash key is present when you are
converting the hash to a string.

Signed-off-by: Yan Liu <yan-liu@ti.com>