PowerHandler: Add extra delay on USB power reset
There are some boards, like beaglebone white, where it is required
to disconnect USB cable in addition to disconnecting power supply
to get a proper reset of the board. Experiments on those boards have
demonstrated that it is possible to get false negatives if there is
not enough time between USB reconnection and the rest of the power
sequence, so add few seconds for boards with multiple power ports to
avoid false negatives.

Signed-off-by: Carlos Hernandez <ceh@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>