SystemLoader: Create weston tsc cal file
Weston expects user to calibrate Touch Screen on first boot. This
creates a problem for automated test as it is not easy to automate
the calibration process. This patch adds a step to create a fake
ws-calibrate.rules file in the SD card boot partition, so the whole
calibration process can be skipped.

Signed-off-by: Carlos Hernandez <ceh@ti.com>