Updated upload_file to go to the tc specific directory
The upload_file method was uploading files one level below the test
case specific directory which causes problems when running multiple
test cases at a time. This issue is that each subsequent test will
overwrite the previous test case's results.

The fix was to use the full path contained in @files_dir which gives
the full path to the test case's specific directory.

Signed-off-by: Calvin Wagner <cwagner@ti.com>
Signed-off-by: Carlos Hernandez <ceh@ti.com>