system_loader: Bring eth0 up first, if using dhcp
NFS server is sensitive to the Ethernet interface the server process runs
on.

Usually the interface that gets the first response from the DHCP server
is used for NFS. If both interfaces are connected to separate networks,
this may cause the boot using NFS as fs to fail occasionally

This patch ensures eth0 always comes up first if both interfaces are
connected to networks with DHCP. The assumption is NFS server is running
on the network that eth0 interface is connected to.

For more info see: https://www.kernel.org/doc/Documentation/filesystems/nfs/nfsroot.txt

Signed-off-by: Elvita Lobo <elvita@ti.com>