Wrap exception msg to avoid marshalling problem
Sometimes when an exception is raised, the object
who raised the exception, instead of the exception
(seems to be a ruby issue) is passed. This causes
vatf to raise a marshal data too short exception
and the original error message is lost. This patch
works around the problem by creating a new exception
object based on the original exception message
before marshalling the test case result objects

Signed-off-by:  Alejandro Hernandeez <ajhernandez@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>