Add connection specific read_for/wait_for funcs
This patch adds two new functions read_on_for and
wait_on_for that are similar to the existing
read_for and wait_for functions but allow the user
to specify the connection in which they want to
read/wait.This enables the user to send commands
in one connection and read/wait on another
connection. The new functions are needed because
the existing read_for and wait_for functions are
hardcoded to use the default connection which is
not guaranteed to be the connection the user would
like to read or wait for a response.

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>