linux_equipment_driver: Add send_sysrq function
In some cases, there is a need to trigger sysrq to provide more
trace to help the developer to debug the problem. Ex, when the
board hangs, it is not sufficient to just have kernel bootlog.
This patch add send_sysrq function in linux_equipment_driver so
user could call it to get more debug traces.

Signed-off-by: Yan Liu <yan-liu@ti.com>