Add exception message to BootloaderException backtrace
This patch changes the BootloaderException class to take
the exception object instead of just the exception
backtrace. Before this patch the exception message was
lost because the backtrace of an exception does not
include the exception's message. After this patch
the exception message is added to the backtrace of
BootloaderException during initialization

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>