Move connection_handler calls inside test case
This patch moves the connection handler related code
inside the test case. With this patch the connection
handler object is instatiated and released within the
test script flow, so problems/changes in one test do
not propagate to the subsequent tests. It also
changes the add_equipment method to optionally pass
euipment_info objects in which case the equipment
class is returned together the with log path. If no
euipment_info object is passed the methods behave as
it did before

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>