atf_session_runner: Allow replacement of equipment obj in add_equipment
This patch adds an additional parameter to the add_equipment method
which allows the user to specify if the equipment added in the
test script can overwrite a previously instantiated object. This
parameter provides flexibility to the user and allows sharing of
resources like a video switch between running scripts. The patch
also includes a change so that equipment logs are also returned
in tests were an exception if thrown, before the equipment logs
created within the test scripts were lost.

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>