Workaround BMC hang issue
It has been noticed that the BMC connection on some boards
hangs after repeated use. To recover from the hang the board
must be power-cycled. Previously for some boot device
scenarios the boards with BMC connections were being rebooted
via the BMC, which eventually caused the BMC connection to
hang on the board. This patch works around that problem by
power-cycling the board regardless if it has a BMC connection
or not

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>