Create driver for K2L EVM
The BMC port in K2L EVMs has an issue that causes it to hang after
some some use. Once the BMC port hangs the board has to be
power cycled to recover the port. This patch creates a new driver
for K2L EVMs in order to workaround the limitation in the BMC port

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>