linux_localhost_driver: Be able to send multiple commands
When sending cmd like "sudo command; echo $?" using the current
driver, right password could not be passed in and failed to
execute the command. This patch is to enable command be passed
in as array like "send_sudo_cmd ["command", "echo $?"] " so command
can be executed correctly. This patch is backward compatible.

Signed-off-by: Yan Liu <yan-liu@ti.com>
Signed-off-by: Carlos Hernandez <ceh@ti.com>