Safeguard against infinite timeout in wait_for
Right now the wait_for method uses any value that is passed
as the timeout. This may hang the system if the value is
zero or nil which causes the Timeout::timeout call to wait
forever. This patch safeguards against such scenarios
by raising an exception if the timeout value is nil,
less or equal zero

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Carlos Hernandez <ceh@ti.com>