KeithleyMultimeter: Add card slot used definition
This patch allows the user to specify the card slot being
used in the Keithley 2700 series multimeter. Before this
patch the command used to specify the channels sampled was
hardcoded to card slot 2, so any multimeter were the card
was connected to slot 1 was not configured correctly and
would return bad data. After this patch the user can set a
'card_slot' => <slot number> hash entry in the params field of
the multimeter entry of the bench and the driver will use that
entry to configure the multimeter

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>