atf_session_runner: Set test_script_found flag sooner
This patch moves the statement to set the flag that saves
the result of loading the test script earlier in the flow.
Before this patch, the flag was being set after calling the
run function of the test script. If an exception was thrown
during the setup or run functions of the test, then the
flag was not set and the clean function of the script was
not called so the cleanup steps defined in the test script
were never executed and the dut may be left in an undesired
state. This patch sets the flag after loading the test script
before calling the setup function of the test script, this
ensures that if an exception is thrown during setup or run
the clean function of the script is also called

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>