lib/sysboot.rb: Refactor get_sysboot_diff
Function get_sysboot_diff would not return the right
length string if the binary representation of the
difference was not the same length as the parameters.
This shortcoming had to be handled in the code calling
get_sysboot_diff. This patch addresses this shortcoming
by formatting the binary representation of the difference.
The patch allowed replacing controller specific calls
in reset_sysboot with a single call to set_sysboot

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>