Implement reboot_regex params in power_cycle
This patch allows the caller to define the regex used
for soft reboot in the power_cycle method. It defines
a new key, 'reboot_regex', for the params Hash of
method power_cycle that when defined overwrites the
default regex used. This gives the caller more control
over the response expected when a reboot command is
run

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Carlos Hernandez <ceh@ti.com>