atf_session_runner: Add month/date in logs path
Currently all logs for tests that have run are store in a directory
under ../system_user folder. If the same parameter are used frequently
over time the number of folders in this directory may reach the maximum
allowed per inode and vatf will fail to run from that point forward until
some maintenance is performed on that folder. This patch adds a new folder
with name <MONTH>_<YEAR> to the logs path so that the limit is not reach
for at least 3 years before requiring maintenance.

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>