Send boot cmd over default serial port
Previously, boot cmd was sent over bootloader port since
U-boot came up on the bootloader port. Now that U-boot comes
up on the main uart, this is no longer necessary.

This change fixes false failures for SoC detection test where
the time to switch between serial ports was enough to cause the
test script to miss caputuring the string it was looking for.

Signed-off-by: Jonah Willis <j-willis@ti.com>