source: Make logs path lower case
The log path folder is based on the host name. If the
host name contain upper case letters the links in the
html page do not work because the letter case will not
match. Since the html page links are lower case make
the logs path lower case to avoid broken links.

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>