target: linux: Refactor uart loading
Functions to load the boards via uart were based on
creating a bash file and then running the file from
the host. This could cause some data loss due to the
time it takes for the file to finish running. This
patch refactors the load functions to load the images
within ruby which provide better control of the load
sequence.

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>